An interface and its implementation were added. The package was given a more accurate name.
