+++ Services/UserServices.ts
@@ -1,0 +1,4 @@
+import { User } from "../user.js";
+export interface UserServices{
+    loadUsers(n: number): Promise<User[]>;
+};
