+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "json-to-html",
+  "name": "Entrega-2",
   "module": "main.ts",
   "type": "module",
   "devDependencies": {
