--- user.json
@@ -1,44 +1,0 @@
-{
-  "results": [
-    {
-      "gender": "female",
-      "name": { "title": "Mrs", "first": "Keira", "last": "Thomas" },
-      "location": {
-        "street": { "number": 9173, "name": "Tweed Street" },
-        "city": "Christchurch",
-        "state": "Southland",
-        "country": "New Zealand",
-        "postcode": 97339,
-        "coordinates": { "latitude": "48.0424", "longitude": "81.2123" },
-        "timezone": { "offset": "+5:45", "description": "Kathmandu" }
-      },
-      "email": "keira.thomas@example.com",
-      "login": {
-        "uuid": "8b9bee1a-f377-401e-8711-2ffea359d80f",
-        "username": "bigdog956",
-        "password": "autumn",
-        "salt": "lDEag6XE",
-        "md5": "84da8bc58910b21de65e8f600d62ee0a",
-        "sha1": "dbd9336ca04ce15fe6917146e3f3ffc659e0a777",
-        "sha256": "823017cba8f1c50b0f4bac69b14402255374a1feb344ff99a8f1bdc04b4a4943"
-      },
-      "dob": { "date": "1999-02-26T03:04:30.306Z", "age": 24 },
-      "registered": { "date": "2020-10-23T04:15:27.351Z", "age": 2 },
-      "phone": "(236)-850-9076",
-      "cell": "(536)-828-7851",
-      "id": { "name": "", "value": null },
-      "picture": {
-        "large": "https://randomuser.me/api/portraits/women/64.jpg",
-        "medium": "https://randomuser.me/api/portraits/med/women/64.jpg",
-        "thumbnail": "https://randomuser.me/api/portraits/thumb/women/64.jpg"
-      },
-      "nat": "NZ"
-    }
-  ],
-  "info": {
-    "seed": "5cb1c03bed2a1099",
-    "results": 1,
-    "page": 1,
-    "version": "1.4"
-  }
-}
