+++ user.ts
@@ -30,13 +30,3 @@ export class User {
     return `${this.name.first} ${this.name.last}`;
   }
 }
-
-export const loadUsers = async (n: number) => {
-  const response = await fetch(`https://randomuser.me/api?results=${n}`);
-  const { results } = (await response.json()) as { results: any[] };
-  const users: Array<User> = [];
-  for (const { gender, name, location, login, email, picture } of results) {
-    users.push(new User(gender, name, location, login, email, picture));
-  }
-  return users;
-};
